moved copyright notice to footer
