+++ README.md
@@ -1,6 +1,6 @@
 ## This is my 11ty site - WIP
 
-See it at [mrnazir.dev](https://www.mrnazir.dev/) 😸
+See it at [imran-nazir.com](https://www.imran-nazir.com/) 😸
 
 [![Netlify Status](https://api.netlify.com/api/v1/badges/711d8724-37d7-4e59-920f-7ca72ce3fdcb/deploy-status)](https://app.netlify.com/sites/jolly-borg-5788ea/deploys)
