+++ src/_includes/footer.njk
@@ -1,0 +1,1 @@
+ <footer class="footer">{% copyrightNotice "imran-nazir.com" %}</footer>
