+++ src/_includes/head.njk
@@ -5,7 +5,7 @@
   <meta http-equiv="X-UA-Compatible" content="ie=edge">
   <link rel="stylesheet" href="/prism.css">
   <link rel="stylesheet" href="/style.css"/>
-  <title>mrnazir.dev</title>
+  <title>imran-nazir.com</title>
   <link rel="stylesheet" href="https://unpkg.com/terminal.css@0.7.2/dist/terminal.min.css"/>
   <script src="https://identity.netlify.com/v1/netlify-identity-widget.js"></script>
 </head>
