+++ src/_includes/nav.njk
@@ -5,6 +5,4 @@
       {% navListItemShortcode entry %}
     {%- endfor %}
   </ul>
-  <hr>
-  &copy; 2021 mrnazir.dev
 </nav>
