+++ src/essays.njk
@@ -3,7 +3,7 @@ layout: layouts/layout.njk
 eleventyNavigation:
   key: Essays
   title: Fragments
-title: mrnazir.dev
+title: imran-nazir.com
 imageSrc: https://images.unsplash.com/photo-1620424393934-04e772be09f4
 imageAlt: 'A man thinking'
 imageAttr: Photo by <a href="https://unsplash.com/@tingeyinjurylawfirm">Tingey Injury Law Firm</a> on <a href="https://unsplash.com/s/photos/thought">Unsplash</a>
