+++ src/index.njk
@@ -1,6 +1,6 @@
 ---
 layout: layouts/layout.njk
-title: mrnazir.dev
+title: imran-nazir.com
 imageSrc: https://images.unsplash.com/photo-1505129039666-fd1461c14ce5
 imageAlt: 'A field of flowers'
 imageAttr: Photo by <a href="https://unsplash.com/@anniespratt">Annie Spratt</a> on <a href="https://unsplash.com/s/photos/kashmir">Unsplash</a>
