+++ src/posts.njk
@@ -3,7 +3,7 @@ layout: layouts/layout.njk
 eleventyNavigation:
   key: Posts
   title: Tech
-title: mrnazir.dev
+title: imran-nazir.com
 imageSrc: https://images.unsplash.com/photo-1510922903530-28ecf3f00362
 imageAlt: 'A notebook on a table'
 imageAttr: Photo by <a href="https://unsplash.com/@bookblock">Bookblock</a> on <a href="https://unsplash.com/s/photos/journal">Unsplash</a>
